+++ Source/ApplicationServices/2Gis.Erm.BLCore.Operations/Generic/Modify/DomainEntityObtainers/DealObtainer.cs
@@ -29,9 +29,9 @@ namespace DoubleGis.Erm.BLCore.Operations.Generic.Modify.DomainEntityObtainers
             entity.Id = dto.Id;
             entity.Name = dto.Name;
             entity.Comment = dto.Comment;
-            entity.StartReason = (int)dto.StartReason;
+            entity.StartReason = dto.StartReason;
             entity.MainFirmId = dto.MainFirmRef.Id;
-            entity.DealStage = (int)dto.DealStage;
+            entity.DealStage = dto.DealStage;
             entity.ClientId = dto.ClientRef.Id.Value;
             entity.OwnerCode = dto.OwnerRef.Id.Value;
